extract & add complex type use case 4 insert on conflict do update set.
