+++ test/typings/test-d/join.test-d.ts
@@ -20,6 +20,7 @@ async function testJoin(db: Kysely<Database>) {
       gender: 'male' | 'female' | 'other'
       modified_at: Date
       marital_status: 'single' | 'married' | 'divorced' | 'widowed' | null
+      deleted_at: Date | null
 
       // Pet columns.
       name: string
@@ -84,6 +85,7 @@ async function testJoin(db: Kysely<Database>) {
     gender: 'male' | 'female' | 'other'
     modified_at: Date
     marital_status: 'single' | 'married' | 'divorced' | 'widowed' | null
+    deleted_at: Date | null
 
     // All Pet columns should be nullable because of the left join
     name: string | null
@@ -111,6 +113,7 @@ async function testJoin(db: Kysely<Database>) {
     gender: 'male' | 'female' | 'other' | null
     modified_at: Date | null
     marital_status: 'single' | 'married' | 'divorced' | 'widowed' | null
+    deleted_at: Date | null
 
     // All Pet columns should also be nullable because there's another
     // right join after the Pet join.
@@ -140,6 +143,7 @@ async function testJoin(db: Kysely<Database>) {
     gender: 'male' | 'female' | 'other' | null
     modified_at: Date | null
     marital_status: 'single' | 'married' | 'divorced' | 'widowed' | null
+    deleted_at: Date | null
 
     name: string | null
     species: 'dog' | 'cat' | null
